requirements.txt
anyio
appnope
argon2-cffi
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
bottleneck
brotli
certifi
cffi
charset-normalizer
comm
debugpy
decorator
defusedxml
exceptiongroup
executing
fastjsonschema
h11
httpcore
httpx
idna
iniconfig
ipykernel
ipython
ipywidgets
isodate
jedi
jinja2
json5
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
jupyterlab_widgets
markupsafe
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numexpr
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
platformdirs
pluggy
ply
prometheus_client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
pygments
pyparsing
PyQt5
PyQt5_sip
PySocks
pytest
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
rdflib
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
Send2Trash
sip
six
sniffio
soupsieve
stack-data
terminado
tinycss2
tomli
tornado
traitlets
typing_extensions
tzdata
urllib3
wcwidth
webencodings
websocket-client
widgetsnbextension
